Move helper methods in Class out from class and into extension class ClassExtensions
