+++ src/AST/Class.cs
@@ -199,128 +199,6 @@ namespace CppSharp.AST
             }
         }
 
-        public IEnumerable<Method> FindOperator(CXXOperatorKind kind)
-        {
-            return Operators.Where(method => method.OperatorKind == kind);
-        }
-
-        public IEnumerable<Method> FindMethodByOriginalName(string name)
-        {
-            return Methods.Where(method => method.OriginalName == name);
-        }
-
-        public IEnumerable<Variable> FindVariableByOriginalName(string originalName)
-        {
-            return Variables.Where(v => v.OriginalName == originalName);
-        }
-
-        public override IEnumerable<Function> GetFunctionOverloads(Function function)
-        {
-            if (function.IsOperator)
-                return FindOperator(function.OperatorKind);
-            return Methods.Where(method => method.Name == function.Name);
-        }
-
-        public IEnumerable<T> FindHierarchy<T>(Func<Class, IEnumerable<T>> func)
-            where T : Declaration
-        {
-            foreach (var elem in func(this))
-                yield return elem;
-
-            foreach (var @base in Bases)
-            {
-                if (!@base.IsClass) continue;
-                foreach(var elem in @base.Class.FindHierarchy<T>(func))
-                    yield return elem;
-            }
-        }
-
-        public Method GetRootBaseMethod(Method @override, bool onlyFirstBase = false)
-        {
-            return (from @base in Bases
-                    where @base.IsClass && (!onlyFirstBase || !@base.Class.IsInterface)
-                    let baseMethod = (
-                        from method in @base.Class.Methods
-                        where
-                            method.Name == @override.Name &&
-                            method.ReturnType == @override.ReturnType &&
-                            method.Parameters.Count == @override.Parameters.Count &&
-                            method.Parameters.SequenceEqual(@override.Parameters,
-                                                            new ParameterTypeComparer())
-                        select method).FirstOrDefault()
-                    let rootBaseMethod = @base.Class.GetRootBaseMethod(@override) ?? baseMethod
-                    where rootBaseMethod != null || onlyFirstBase
-                    select rootBaseMethod).FirstOrDefault();
-        }
-
-        public Property GetRootBaseProperty(Property @override, bool onlyFirstBase = false)
-        {
-            return (from @base in Bases
-                    where (!onlyFirstBase || !@base.Class.IsInterface) && @base.IsClass
-                    let baseProperty = (
-                        from property in @base.Class.Properties
-                        where
-                            property.Name == @override.Name &&
-                            property.Parameters.Count == @override.Parameters.Count &&
-                            property.Parameters.SequenceEqual(@override.Parameters,
-                                                            new ParameterTypeComparer())
-                        select property).FirstOrDefault()
-                    let rootBaseProperty = @base.Class.GetRootBaseProperty(@override) ?? baseProperty
-                    where rootBaseProperty != null || onlyFirstBase
-                    select rootBaseProperty).FirstOrDefault();
-        }
-
-        public Property GetPropertyByName(string propertyName)
-        {
-            Property property = Properties.FirstOrDefault(m => m.Name == propertyName);
-            if (property != null)
-                return property;
-            Declaration decl;
-            foreach (var baseClassSpecifier in Bases.Where(
-                b => b.Type.IsTagDecl(out decl) && !b.Class.Ignore))
-            {
-                property = baseClassSpecifier.Class.GetPropertyByName(propertyName);
-                if (property != null)
-                    return property;
-            }
-            return null;
-        }
-
-        public Property GetPropertyByConstituentMethod(Method method)
-        {
-            var property = Properties.FirstOrDefault(p => p.GetMethod == method);
-            if (property != null)
-                return property;
-            property = Properties.FirstOrDefault(p => p.SetMethod == method);
-            if (property != null)
-                return property;
-            Declaration decl;
-            foreach (BaseClassSpecifier @base in Bases.Where(b => b.Type.IsTagDecl(out decl)))
-            {
-                property = @base.Class.GetPropertyByConstituentMethod(method);
-                if (property != null)
-                    return property;
-            }
-            return null;
-        }
-
-        public Method GetMethodByName(string methodName)
-        {
-            var method = Methods.FirstOrDefault(
-                // HACK: because of the non-shared v-table entries bug one copy may have been renamed and the other not
-                m => string.Compare(m.Name, methodName, StringComparison.OrdinalIgnoreCase) == 0);
-            if (method != null)
-                return method;
-            Declaration decl;
-            foreach (BaseClassSpecifier @base in Bases.Where(b => b.Type.IsTagDecl(out decl)))
-            {
-                method = @base.Class.GetMethodByName(methodName);
-                if (method != null)
-                    return method;
-            }
-            return null;
-        }
-
         public override T Visit<T>(IDeclVisitor<T> visitor)
         {
             return visitor.VisitClassDecl(this);
